4.75 CEX / Universal COBRA 7.1
This is a universal project, meaning this supports all current public
cfw releases except those cfws built with mfwbuilders.

In order for those cfws to work with this source, those 64bit hashes are
required to be ported in

modulespatch.h/c due to its hash check preventing unknown sprx/self
modules to be patched on boot.

as a result it can cause issues with ps2/psp games when this cobra
payload is applied without proper porting

Currently following CFWs are supported.

1. mine [unreleased]
2. HABIB 4.75 Starbucks 1.00
3. Ferrorx 4.75 1.00 [Untested but it should work]